requirements.txt
anthropic>=0.25.0
pandas>=2.0.0
flask>=3.0.0
python-dotenv>=1.0.0
